Automatically add and remove transform-sets.
Until now only references to transform-sets were added to crypto maps.
Previously, the corresponding transform-set defintion had to be added
manually.

Added new attribute {simple_object} in $structure.  With this
attribute, an objects isn't changed, but always transferred to device.
But before transfer, all objects on device are checked, if an object
with identical attribute / value pairs is already available. If found,
this object is used and no transfer occurs.