No longer move ACL line for device access
This could occurr, if a deny rule was inserted into a block of permit rules
or if a deny rule was removed from a block of permit rules.

Blocks of ACL lines are no longer sorted.
If an ACL line is moved inside the same block,
this move operation is simply discarded now.